% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{are_waedenswil}
\alias{are_waedenswil}
\title{Public Transport connection quality in the city of Wädenswil}
\format{
Simple feature collection 13 features and 5 fields
\describe{
  \item{KLASSE}{Transport connection categorised in 4 main classes}


}
}
\source{
\url{https://s.geo.admin.ch/7e80a8bd28}
plot(sf::st_geometry(are_waedenswil))
}
\usage{
are_waedenswil
}
\description{
Public Transport connection quality in the city of Wädenswil
}
\keyword{datasets}
