% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bezirke}
\alias{bezirke}
\title{Bezirke}
\format{
Simple feature collection with 192 features and 4 fields
\describe{
  \item{NAME}{Name of every canton and as well as it's respective districts (Bezirke)}
  \item{OBJEKTART}{It differentiates the depicted polygon features into Cantons (Kanton) and Districts (Bezirke)}
}
}
\source{
\urlhttps {https://map.geo.admin.ch/}
}
\usage{
bezirke
}
\description{
The dataset depicts the districts (Bezirke) of every canton of the country of Switzerland.
}
\keyword{datasets}
