% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chessboard}
\alias{chessboard}
\title{Chessboard}
\format{
\describe{
  \item{colour}{The colour of the field}
  \item{namex}{}
  \item{namey}{}
  \item{name}{The name of the field}
  \item{namex}{The integer of the field}
}
}
\usage{
chessboard
}
\description{
A chessboard in an `sf` format with 64 features and 5 fields, generated from
scratch
}
\keyword{datasets}
