% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eire}
\alias{eire}
\title{Eire}
\format{
Simple feature collection with 26 features and 10 fields
\describe{
  \item{names}{Name of each of the county}
  \item{INCOME}{Average income of the respective county}
}
}
\source{
\urlhttps {https://cran.r-project.org/web/packages/spData/index.html}
}
\usage{
eire
}
\description{
The dataset depicts the counties of the country of Ireland. The counties of Ireland are historic administrative divisions of the island.
}
\keyword{datasets}
