% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gemeinde_zh}
\alias{gemeinde_zh}
\title{Municipalities (Gemeinden) in the canton of Zurich}
\format{
Simple feature collection with 166 features and 7 fields
\describe{
  \item{gemeindena}{Name of the Municipality (Gemeinde)}
  \item{bezirksna}{Name of the District (Bezirk)}
}
}
\source{
\urlhttps {https://maps.zh.ch/}
}
\usage{
gemeinde_zh
}
\description{
The dataset depicts all the municipalities (Gemeinden) in the canton of Zurich
}
\keyword{datasets}
