% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gemeindegrenzen_zh}
\alias{gemeindegrenzen_zh}
\title{Gemeindegrenzen im Kanton Zürich}
\format{
A sf object with 167 features and 10 variables
\describe{
  \item{BFS}{BFS Number}
  \item{ARPS}{Add description here}
  \item{BEZIRKSNAME}{Add description here}
  \item{ART_TEXT}{Add description here}
  \item{GEMEINDENAME}{Add description here}
  \item{AREA_ROUND}{Add description here}
  \item{SHAPE_AREA}{Add description here}
  \item{SHAPE_LEN}{Add description here}
  \item{geom}{Add description here}
}
}
\source{
\url{http://maps.zh.ch/?topic=ortho19ZH&showtab=ogddownload}
}
\usage{
gemeindegrenzen_zh
}
\description{
This dataset contains the boarder for the municipalities in the canton of Zurich
The data was obtained from maps.zh.ch
}
\keyword{datasets}
