% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haltestelle_bahn}
\alias{haltestelle_bahn}
\title{Train stations locations in Switzerland}
\format{
Simple feature collection with 3134 features and 16 fields
\describe{
  \item{name}{Name of the train station}


}
}
\source{
\url{https://map.geo.admin.ch}
}
\usage{
haltestelle_bahn
}
\description{
The dataset contains a collection of point features that represent
the locations of the train stations across the country of Switzerland
}
\examples{
#plot(sf::st_geometry(haltestelle_bahn))
}
\keyword{datasets}
