% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo}
\alias{meteo}
\title{Meteorological Data from Weather Stations in Switzerland}
\format{
A tibble object with 1262615 features and 3 variables
\describe{
  \item{stn}{The Station where the temperature was measured}
  \item{time}{time in the format YYYYMMDDHH}
  \item{tre200h0}{Temperature in degrees celcius}
}
}
\source{
\url{https://gate.meteoswiss.ch/idaweb/}
}
\usage{
meteo
}
\description{
The dataset contains temperature data ranging from 2000 to 2005
TODO: Add more description here:
}
\details{
- what is the interval?
- where was the temperature measured
- add description of _legend
}
\keyword{datasets}
