% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mountainBikes_routes}
\alias{mountainBikes_routes}
\title{Mountain bike routes in Switzerland}
\format{
Simple feature collection with 286 features and 29 fields
\describe{
  \item{beschreib}{Description of the mountain bike route - Starting point to End point}


}
}
\source{
\url{https://map.geo.admin.ch}
}
\usage{
mountainBikes_routes
}
\description{
The dataset contains a line spatial feature that depicts all the mountain bike routes in Switzerland
}
\keyword{datasets}
