% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_india2010}
\alias{pop_india2010}
\title{India's Population in 2010}
\format{
RasterLayer with dimensions : 690, 701, 483690  (nrow, ncol, ncell)
\describe{
  \item{}{}
}
}
\source{
\url{https://sedac.ciesin.columbia.edu/data/collection/gpw-v4}
}
\usage{
pop_india2010
}
\description{
The dataset depicts India’s population at 2010 timestamp
}
\keyword{datasets}
