% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recycling_raster}
\alias{recycling_raster}
\title{Recycling_raster}
\format{
RasterLayer with dimensions : 228, 222, 50616  (nrow, ncol, ncell) and resolution : 25, 25  (x, y)
\describe{
  \item{}{}
  \item{}{}
}
}
\source{
\url{https://modul-agi.github.io/HS20/}
}
\usage{
recycling_raster
}
\description{
The dataset depicts the recycling facilities in the city of Wädenswil (Canton of Zurich, Switzerland).
}
\keyword{datasets}
