% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smr50_kgrs_mosaic}
\alias{smr50_kgrs_mosaic}
\title{SMR50_kgrs_mosaic}
\format{
\describe{
  \item{}{}
  \item{}{}
}
}
\source{
\urlhttps {https://map.geo.admin.ch/}
}
\usage{
smr50_kgrs_mosaic
}
\description{
SMR50_kgrs_mosaic
}
\keyword{datasets}
