% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smr50_krel_mosaic}
\alias{smr50_krel_mosaic}
\title{SMR50_krel_mosaic}
\format{
\describe{
  \item{}{}
  \item{}{}
}
}
\source{
\url{https://map.geo.admin.ch/}
}
\usage{
smr50_krel_mosaic
}
\description{
SMR50_krel_mosaic
}
\keyword{datasets}
