% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swissAlti3D}
\alias{swissAlti3D}
\title{SwissAlti3D}
\format{
RasterLayer with resolution: 10, 10  (x, y) and dimensions: 100, 100, 10000  (nrow, ncol, ncell)
}
\source{
\urlhttps {https://cms.geo.admin.ch/Topo/swissalti3d/swissalti3dgeotifflv95.zip}
}
\usage{
swissAlti3D
}
\description{
The dataset depicts the digital elevation model for the extent of the country of Switzerland
}
\keyword{datasets}
