% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zweitwohnung_gemeinden}
\alias{zweitwohnung_gemeinden}
\title{Zweitwohnung_Gemeinden}
\format{
Simple feature collection with 2354 features and 14 fields
\describe{
  \item{gemeinde}{Name of the Municipality (Gemeinde)}
  \item{kanton}{Name of the Canton to which the municipality belongs}
  \item{bezirk}{Name of the district (Bezirk)}
}
}
\source{
\url{https://map.geo.admin.ch}
}
\usage{
zweitwohnung_gemeinden
}
\description{
The dataset depicts all the municipalities (Gemeinden) with the respective information about the Second Home ..
}
\keyword{datasets}
