% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{badeplaetze_zh}
\alias{badeplaetze_zh}
\title{Swimming spots (Badenplätze) in the Canton of Zurich}
\format{
Simple feature collection 71 features and 6 fields
\describe{
  \item{x_coord}{X coordinate in CH1903+LV95}
  \item{y_coord}{Y coordinate in CH1903+LV95}
  \item{gewaesser}{Water "body" name near to which the "spot lies"}
}
}
\source{
\url{https://opendata.swiss/de/dataset/badeplatze1}
}
\usage{
badeplaetze_zh
}
\description{
The dataset contains 71 swimming spots within the Canton of Zürich
}
\examples{
plot(sf::st_geometry(badeplaetze_zh))
}
\keyword{datasets}
