% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kantonsgebiet}
\alias{kantonsgebiet}
\title{Kantonsgebiet}
\format{
Simple feature collection with 56 features and 22 fields
\describe{
  \item{name}{Name of every Canton in Switzerland}
}
}
\source{
\urlhttps {https://shop.swisstopo.admin.ch/de/products/landscape/boundaries3D}
}
\usage{
kantonsgebiet
}
\description{
Kantonsgebiet
}
\keyword{datasets}
